Add 'loadText' method for loading large text files
